initial reorganising
